LEVEL SELECTOR
omg i spent so long dicking with the ui. but um, it works, you gotta
open up build settings and add the scenes in and stuff, then set the
scene index in the level selection, which is on each panel. Still needs
transitions. We also still need to do win screens.